move NewsletterResponse.ts to /src/lib
