+++ dotcom-rendering/src/model/NewsletterResponse.ts
