+++ dotcom-rendering/src/web/components/EmailSignup.tsx
@@ -6,7 +6,7 @@ import {
 	space,
 	textSans,
 } from '@guardian/source-foundations';
-import type { NewsletterResponse } from 'NewsletterResponse';
+import type { NewsletterResponse } from 'src/model/NewsletterResponse';
 import { NewsletterFrequency } from './NewsletterFrequency';
 import { SecureSignup } from './SecureSignup';
