+++ dotcom-rendering/src/web/lib/ArticleRenderer.tsx
@@ -2,7 +2,7 @@ import { css } from '@emotion/react';
 import type { ArticleFormat } from '@guardian/libs';
 import { ArticleDesign } from '@guardian/libs';
 import { from } from '@guardian/source-foundations';
-import type { NewsletterResponse } from 'NewsletterResponse';
+import type { NewsletterResponse } from 'src/model/NewsletterResponse';
 import {
 	adCollapseStyles,
 	labelStyles as adLabelStyles,
